+++ src/Symfony/Component/PropertyAccess/Tests/DependencyInjection/Fixtures/property_accessor_overridden.php
@@ -1,0 +1,10 @@
+<?php
+
+$container->loadFromExtension('property_access', [
+    'magic_call' => true,
+    'magic_get' => true,
+    'magic_set' => false,
+    'throw_exception_on_invalid_index' => true,
+    'throw_exception_on_invalid_property_path' => false,
+    'wildcard_reads' => true,
+]);