+++ src/Symfony/Component/PropertyAccess/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add `PropertyAccessBundle`, which provides the `property_access` configuration and the services previously provided by `FrameworkBundle` under `framework.property_access`
  * Add `PropertyAccessorBuilder::enableWildcardReads()` to read every element of a collection through a `[*]` path
 
 8.1